requirements.txt
# Core requirements that also requires updating in setup.py
requests==2.18.4
PyYaml==3.12

# Developer requirements
